Do not early return on null `bucket_ptr`
Buckets are allocated on demand based on `Thread::bucket`.
This means that when only threads with a high `id` (and thus high `bucket`) are writing entries into the `ThreadLocal`,
only higher `buckets` will be allocated, and lower buckets will be `null`.

Thus we must not early-return when encounting a `null` bucket.